backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Jinja2==3.1.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Werkzeug==2.3.4